Add the membind example program.
This replaces all the membind-i tests into something more portable and easy to use. 
The purpose of this test is to experiment dynamic threads and data placements.
The user can specify how many threads will be created, where will they be scheduled from,
which mbind policy to use on the accessed data, and where to distribute it.


git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@19544 ee107e0b-c2fa-0310-842a-ec2723fb34f4