Modify the memteams example program so as to be able to specify on which NUMA node each array will be bound to.
You can pass a list of integers on the command line to express which NUMA node will hold which array.

As an example, "pm2-load memteams local bind 3 3 0 0" will bind the first two arrays on node 3, and the last two on node 0, on a 4-nodes computer.

For now, passing 'interleave' as memory allocation mode means "interleave on every node of the computer". That's why passing a list of nodes with the 'interleave' option is useless (i.e.: the passed node list will be ignored).



git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@19927 ee107e0b-c2fa-0310-842a-ec2723fb34f4