Bubble scheduler does not have to act weird anymore about whether they are distributing threads and bubbles for the first time or not.
This patch modifies the behavior of the `submit' function 
of the cache and memory bubble schedulers, so that they 
actually call their specific distribution algorithm each time 
somebody calls them.

This modifies the behavior of the marcel_bubble_begin () function,
which acts now as the marcel_bubble_shake () function, but remains 
available in Marcel to keep applications and libraries that use 
this function working.

Note that it's now up to the programmer to control how many times
marcel_bubble_sched_begin () will be called to distributes the threads 
and bubbles of his application.

(Longest log message ever (TM))



git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@20248 ee107e0b-c2fa-0310-842a-ec2723fb34f4