Marcel: Add `--marcel-topology-fsys-root' option.
This allows the file system root to be explicitly specified on GNU-based
systems, which can be used to tell the topology discovery code to use,
an alternate root directory when opening `/sys' files.  This makes it
possible to write unit tests for the topology code without having to use
`chroot(1)'.

git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@20584 ee107e0b-c2fa-0310-842a-ec2723fb34f4