Marcel: Use PukABI when building Marcel's `libpthread.so'.
This supposedly fixes problems such as libc function calls preempted by
Marcel, and called again by another Marcel user-level thread in the same
LWP, which glibc's low-level lock cannot handle (one typical example of
this is with malloc(3)).  PukABI intercepts malloc(3) and other glibc
functions and guarantees that they are not called "at the same time" by
several threads of the same LWP.

The resulting `libpthread.so' just has to be `LD_PRELOAD'ed the usual
way.

git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@20662 ee107e0b-c2fa-0310-842a-ec2723fb34f4