On Linux kernel >= 2.6.22, the numa node of each device may be read directly from /sys/.../numa_node. It works and it's easy.
On older kernels, the corresponding cpumask is available in
/sys/.../local_cpus but libnuma2 does not provide any easy way
to find the corresponding numa node. We could write some ugly
code to do it, but we don't really care about old kernels
for NUIOA anyway. So we disable this code, and we just use 
the 'numa_node' sysfs attribute when it exists.


git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@21396 ee107e0b-c2fa-0310-842a-ec2723fb34f4